Fixed issue #112 - Deep JavaScript Object traversal not working in worker threads
The actual issue is that pass-by-reference functions seem to not work correctly when
called across threads with CallOnMainThread. This was an issue with GetMemberNames
which passed a vector by reference to be filled out. I created a version of the
function that took a pointer and made it use that when making cross-thread calls
and it should now work correctly.

Conflicts:

	examples/FBTestPlugin/FBTestPluginAPI.h
	src/ScriptingCore/JSAPI.cpp