Stable eventId for NPObjectAPI.
Safari rewraps a Javascript function for each call. This means that
removeEventListener will never remove a listener set by
addEventListener. Use a sequence to set the eventId in the __fb_event_id
property on the Javascript object so that we'll be able to see it when
the object is passed in. Unfortunately, this requires removing the const
flag on getEventId.