Fixed an exception slicing bug described in the question: http://stackoverflow.com/questions/33647949/firebreath-promise-reject-slicing-exception.
This bug appears only in Mac OS X and Linux builds. The bug is caused by the implementations of std::exception in these OS.
Its copying constructors instead of copying std::exception::whats value use default value std::Exception.

The bug was fixed by moving FB::Deferred and FB::Promise classes on using std::exception_ptr
intended to pass pointer to an exception out of a function that caught it and re-throw in a different thread.